feat: silent save telegraph
