+++ client/bot/handlers/utils/shortcut/tftask.go
@@ -1,4 +1,3 @@
-// Some shortcuts for duplicate code in handlers, they should return dispatcher errors
 package shortcut
 
 import (
@@ -7,14 +6,10 @@ import (
 
 	"github.com/celestix/gotgproto/dispatcher"
 	"github.com/celestix/gotgproto/ext"
-	"github.com/celestix/gotgproto/types"
 	"github.com/charmbracelet/log"
 	"github.com/gotd/td/tg"
-	"github.com/krau/SaveAny-Bot/client/bot/handlers/utils/mediautil"
 	"github.com/krau/SaveAny-Bot/client/bot/handlers/utils/msgelem"
-	"github.com/krau/SaveAny-Bot/client/bot/handlers/utils/re"
 	"github.com/krau/SaveAny-Bot/client/bot/handlers/utils/ruleutil"
-	"github.com/krau/SaveAny-Bot/common/cache"
 	"github.com/krau/SaveAny-Bot/common/utils/tgutil"
 	"github.com/krau/SaveAny-Bot/core"
 	"github.com/krau/SaveAny-Bot/core/batchtftask"
@@ -24,40 +19,6 @@ import (
 	"github.com/krau/SaveAny-Bot/storage"
 	"github.com/rs/xid"
 )
-
-// 获取消息中的文件并回复等待消息, 返回等待消息, 获取到的文件
-func GetFileFromMessageWithReply(ctx *ext.Context, update *ext.Update, message *tg.Message, tfileopts ...tfile.TGFileOptions) (replied *types.Message,
-	file tfile.TGFileMessage, err error,
-) {
-	logger := log.FromContext(ctx)
-	media := message.Media
-	supported := mediautil.IsSupported(media)
-	if !supported {
-		ctx.Reply(update, ext.ReplyTextString("不支持的消息类型"), nil)
-		return nil, nil, dispatcher.EndGroups
-	}
-
-	replied, err = ctx.Reply(update, ext.ReplyTextString("正在获取文件信息..."), nil)
-	if err != nil {
-		logger.Errorf("Failed to reply: %s", err)
-		return nil, nil, dispatcher.EndGroups
-	}
-	options := []tfile.TGFileOptions{
-		tfile.WithMessage(message),
-	}
-	if len(tfileopts) > 0 {
-		options = append(options, tfileopts...)
-	} else {
-		options = append(options, tfile.WithNameIfEmpty(tgutil.GenFileNameFromMessage(*message)))
-	}
-	file, err = tfile.FromMediaMessage(media, message, options...)
-	if err != nil {
-		logger.Errorf("Failed to get file from media: %s", err)
-		ctx.Reply(update, ext.ReplyTextString("获取文件失败: "+err.Error()), nil)
-		return nil, nil, dispatcher.EndGroups
-	}
-	return replied, file, nil
-}
 
 // 创建一个 tftask.TGFileTask 并添加到任务队列中, 以编辑消息的方式反馈结果
 func CreateAndAddTGFileTaskWithEdit(ctx *ext.Context, userID int64, stor storage.Storage, dirPath string, file tfile.TGFileMessage, trackMsgID int) error {
@@ -118,74 +79,6 @@ func CreateAndAddTGFileTaskWithEdit(ctx *ext.Context, userID int64, stor storage
 	})
 
 	return dispatcher.EndGroups
-}
-
-type EditMessageFunc func(text string, markup tg.ReplyMarkupClass)
-
-// 获取链接中的文件并回复等待消息
-func GetFilesFromUpdateLinkMessageWithReplyEdit(ctx *ext.Context, update *ext.Update) (replied *types.Message, files []tfile.TGFileMessage, editReplied EditMessageFunc, err error) {
-	logger := log.FromContext(ctx)
-	msgLinks := re.TgMessageLinkRegexp.FindAllString(update.EffectiveMessage.GetMessage(), -1)
-	if len(msgLinks) == 0 {
-		logger.Warn("no matched message links but called handleMessageLink")
-		return nil, nil, nil, dispatcher.EndGroups
-	}
-	replied, err = ctx.Reply(update, ext.ReplyTextString("正在获取消息..."), nil)
-	if err != nil {
-		logger.Errorf("failed to reply: %s", err)
-		return nil, nil, nil, dispatcher.EndGroups
-	}
-	editReplied = func(text string, markup tg.ReplyMarkupClass) {
-		if _, err := ctx.EditMessage(update.EffectiveChat().GetID(), &tg.MessagesEditMessageRequest{
-			ID:          replied.ID,
-			Message:     text,
-			ReplyMarkup: markup,
-		}); err != nil {
-			logger.Errorf("failed to edit message: %s", err)
-		}
-	}
-
-	files = make([]tfile.TGFileMessage, 0, len(msgLinks))
-	for _, link := range msgLinks {
-		chatId, msgId, err := tgutil.ParseMessageLink(ctx, link)
-		if err != nil {
-			logger.Errorf("failed to parse message link %s: %s", link, err)
-			continue
-		}
-		msg, err := tgutil.GetMessageByID(ctx, chatId, msgId)
-		if err != nil {
-			logger.Errorf("failed to get message by ID: %s", err)
-			continue
-		}
-		media, ok := msg.GetMedia()
-		if !ok {
-			logger.Debugf("message %d has no media", msg.GetID())
-			continue
-		}
-		file, err := tfile.FromMediaMessage(media, msg, tfile.WithNameIfEmpty(tgutil.GenFileNameFromMessage(*msg)))
-		if err != nil {
-			logger.Errorf("failed to create file from media: %s", err)
-			continue
-		}
-		files = append(files, file)
-	}
-	if len(files) == 0 {
-		editReplied("没有找到可保存的文件", nil)
-		return nil, nil, nil, dispatcher.EndGroups
-	}
-	return replied, files, editReplied, nil
-}
-
-func GetCallbackDataWithAnswer[DataType any](ctx *ext.Context, update *ext.Update, dataid string) (DataType, error) {
-	data, ok := cache.Get[DataType](dataid)
-	if !ok {
-		log.FromContext(ctx).Warnf("Invalid data ID: %s", dataid)
-		queryID := update.CallbackQuery.GetQueryID()
-		ctx.AnswerCallback(msgelem.AlertCallbackAnswer(queryID, "数据已过期或无效"))
-		var zero DataType
-		return zero, dispatcher.EndGroups
-	}
-	return data, nil
 }
 
 // 创建一个 batchtftask.BatchTGFileTask 并添加到任务队列中, 以编辑消息的方式反馈结果
